Venkatesh : update in email controller change static values to dynamic values and push notifications create in separate folder
